+++ src/app/accounts/register/register.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {AbstractControl, FormGroup} from '@angular/forms';
 import {environment} from '@environments/environment';
-import {RegisterService} from '@app/_services/api/accounts/register.service';
+import {RegisterService} from '@app/accounts/_services/register.service';
 import {ToastrService} from "ngx-toastr";
 import {RegisterForm} from "@app/accounts/_forms/register.form";
